US8 : Added invite method

+++ src/main/java/com/hackathon/event/dto/ParticipantRequestDto.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class ParticipantRequestDto {
     private String email;
-    private Integer registrationId;
+    private Long registrationId;
 }
